requirements.txt
numpy 
pandas 
openpyxl 
tqdm 
gensim 
matplotlib 
seaborn 
plotly 
umap-learn